test : 추가된 기능 및 리팩토링에 따른 테스트 코드 작성
