+++ src/test/java/tback/kicketingback/user/signup/SignupTest.java
@@ -26,7 +26,6 @@ import tback.kicketingback.user.signup.service.DefaultSignUpService;
 
 @SpringBootTest
 public class SignupTest {
-
 	@Autowired
 	private EmailAuthService emailAuthService;
